refactor(balancer) namespaced files and refactored handlers
The balancer related code now lives under /balancer. And the
handlers 'onXxx', 'beforeYyy', etc. now have empty implementations
in the base class, so no more need to call the 'super'.
