+++ spec/balancer/base_spec.lua
@@ -16,9 +16,9 @@ describe("[balancer_base]", function()
   local snapshot
 
   setup(function()
-    _G.package.loaded["dns.client"] = nil -- make sure module is reloaded
+    _G.package.loaded["resty.dns.client"] = nil -- make sure module is reloaded
     _G._TEST = true  -- expose internals for test purposes
-    balancer_base = require "resty.dns.balancer_base"
+    balancer_base = require "resty.dns.balancer.base"
     client = require "resty.dns.client"
   end)
