+++ src/resty/dns/balancer/least_connections.lua
@@ -14,7 +14,7 @@
 -- @license Apache 2.0
 
 
-local balancer_base = require "resty.dns.balancer_base"
+local balancer_base = require "resty.dns.balancer.base"
 local binaryHeap = require "binaryheap"
 local ngx_log = ngx.log
 local ngx_DEBUG = ngx.DEBUG
@@ -101,7 +101,6 @@ end
 -- removing the address, so delete from binaryHeap
 function lc:onRemoveAddress(address)
   self.binaryHeap:remove(address)
-  self.super.onRemoveAddress(self, address)
 end
